refactor(config): split ClientConfig/ServerConfig into generic + transport-specific modules
Make the config layer transport-agnostic so callers go through a single
polymorphic entry point.

Client side:
- core/config/src/client.rs now owns only the transport-agnostic
  ClientConfig struct, builders, validation, Display, KeepaliveConfig,
  AuthenticationConfig, BackoffConfig, TransportChannel<G>, and the
  public dispatch method to_channel().
- core/config/src/grpc/client.rs owns all gRPC-specific impl methods
  (channel construction, proxy tunnels, Unix sockets, auth layer
  assembly) under a crate-private to_grpc_channel().
- core/config/src/websocket/client.rs to_websocket_channel() is now
  pub(crate); to_channel() is the only public entry point.

Server side:
- core/config/src/server.rs owns only the transport-agnostic
  ServerConfig struct, builders, validation, Display,
  KeepaliveServerParameters, AuthenticationConfig.
- core/config/src/grpc/server.rs owns the gRPC server methods (renamed
  to_grpc_server_future -> to_server_future, run_grpc_server ->
  run_server) plus tonic builder/incoming/Unix plumbing.
- WebSocket server keeps run_websocket_server; the datapath
  polymorphically dispatches on config.transport.

Backward-compat re-exports under grpc::client and grpc::server keep
existing import paths (used by the rust bindings) working.

External callers updated:
- core/controller/src/service.rs: use to_channel() + TransportChannel
  pattern match.
- slimctl/src/client.rs: same pattern; bail on Websocket variant.
- core/config/tests/e2e.rs: pattern match on TransportChannel::Grpc.
